chore: move ipfs utils to JSON utils
